+++ dces-drc-integration/src/main/java/uk/gov/justice/laa/crime/dces/integration/client/DrcClient.java
@@ -3,8 +3,8 @@ package uk.gov.justice.laa.crime.dces.integration.client;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.service.annotation.HttpExchange;
 import org.springframework.web.service.annotation.PostExchange;
-import uk.gov.justice.laa.crime.dces.integration.model.drc.SendContributionFileDataToDrcRequest;
-import uk.gov.justice.laa.crime.dces.integration.model.drc.SendFdcFileDataToDrcRequest;
+import uk.gov.justice.laa.crime.dces.integration.model.SendContributionFileDataToDrcRequest;
+import uk.gov.justice.laa.crime.dces.integration.model.SendFdcFileDataToDrcRequest;
 
 @HttpExchange("/drc-client")
 public interface DrcClient {
