+++ dces-drc-integration/src/main/java/uk/gov/justice/laa/crime/dces/integration/model/SendContributionFileDataToDrcRequest.java
@@ -1,4 +1,4 @@
-package uk.gov.justice.laa.crime.dces.integration.model.drc;
+package uk.gov.justice.laa.crime.dces.integration.model;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
