+++ dces-drc-integration/src/main/java/uk/gov/justice/laa/crime/dces/integration/model/exception/ErrorResponse.java
@@ -1,4 +1,4 @@
-package uk.gov.justice.laa.crime.dces.integration.rest.common;
+package uk.gov.justice.laa.crime.dces.integration.model.exception;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
