+++ dces-drc-integration/src/main/java/uk/gov/justice/laa/crime/dces/integration/model/exception/context/ErrorResponseBodyWrapper.java
@@ -1,4 +1,4 @@
-package uk.gov.justice.laa.crime.dces.integration.configuration.wrapper;
+package uk.gov.justice.laa.crime.dces.integration.model.exception.context;
 
 import lombok.AllArgsConstructor;
 import org.springframework.core.MethodParameter;
@@ -8,7 +8,6 @@ import org.springframework.http.server.ServerHttpRequest;
 import org.springframework.http.server.ServerHttpResponse;
 import org.springframework.web.bind.annotation.ControllerAdvice;
 import org.springframework.web.servlet.mvc.method.annotation.ResponseBodyAdvice;
-import uk.gov.justice.laa.crime.dces.integration.model.exception.TraceData;
 import uk.gov.justice.laa.crime.dces.integration.service.TraceService;
 
 import java.lang.reflect.Method;
