+++ dces-drc-integration/src/main/java/uk/gov/justice/laa/crime/dces/integration/model/exception/context/TraceData.java
@@ -1,0 +1,18 @@
+package uk.gov.justice.laa.crime.dces.integration.model.exception.context;
+
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+import java.sql.Timestamp;
+
+@Builder
+@Data
+@NoArgsConstructor
+@AllArgsConstructor
+public class TraceData {
+
+    private String traceId;
+    private Timestamp timestamp;
+}
